requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.41.0
requests==2.32.5
poetry==2.3.2
python-dotenv==1.2.1
aiogram==3.25.0
aiohttp==3.13.3
pymongo==4.12.1
psycopg2-binary==2.9.10
